refactor(cards): merge bonuses/penalties into one component (QUAL-2)
The two cards were ~95% identical (~1140 lines each); every fix (e.g.
the ERR-5 double-submit guard) had to be made twice. Extract one
parameterised component taskmate-incentive-card.js (createIncentiveCard(P))
generated faithfully from the bonuses card with every kind-specific value
parameterised: tag, kind (service prefix add_/apply_/update_/remove_),
idKey, i18n prefix (via _tp helper), list attribute, icons (form/header/
apply), points sign (classic + designed), accent colours (injected as CSS
vars at render so css stays interpolation-free), designed accent/hd/
icons, card name/desc, banner.

bonuses-card.js and penalties-card.js are now ~30-line ES-module wrappers
that import the factory and pass their P. Net -1090 lines.

Verification: zero stray kind-specific literals in the factory (grep);
every P value cross-checked against the originals; structural checks
(2 _tp defs, 4 parameterised services, 3 idKey sites, pv wiring, P.sign);
node --check + ESLint clean; served + cross-module import resolve on
ha-dev; browserless load test: both cards + both editors register with
no page errors.
